Move homeassistant_proxy/ under homeassistant/proxy/
Group HA proxy with other homeassistant components. Updates Python
imports, Bazel targets, and CI references.
