+++ homeassistant/proxy/conftest.py
@@ -1,4 +1,4 @@
-"""pytest-asyncio auto mode for homeassistant_proxy tests."""
+"""pytest-asyncio auto mode for homeassistant.proxy tests."""
 
 
 def pytest_configure(config):
